Move iteration section to react component
