+++ app/javascript/components/common/index.ts
@@ -9,3 +9,4 @@ export { Pagination } from './Pagination'
 export { Reputation } from './Reputation'
 export { Avatar } from './Avatar'
 export { Modal } from './Modal'
+export { ProminentLink } from './ProminentLink'
