--- app/views/tracks/exercises/show/_iteration_section.html.haml
@@ -1,31 +1,0 @@
-- unless iteration
-  - return
-
-%section.latest-iteration
-  = ReactComponents::Student::SolutionSummary.new(iteration)
-
-  = link_to Exercism::Routes.track_exercise_iteration_path(track, exercise, iteration), class: "iteration" do
-    = ReactComponents::Track::IterationSummary.new(iteration)
-    = graphical_icon 'chevron-right', css_class: 'action-icon'
-
-  = render ViewComponents::ProminentLink.new("See all of your iterations", track_concepts_path(track))
-
-  .next-steps
-    .community-solutions
-      = graphical_icon "graphic-community-solutions", css_class: "header-icon"
-      %h3 Learn from others' solutions
-      %p Explore their approaches to learn new tips and tricks. Discover popular solutions to this exercise.
-      = link_to "View community solutions", "#", class: 'btn-small-discourage'
-
-    .mentoring
-      = graphical_icon "graphic-mentoring-screen", css_class: "header-icon"
-      %h3 Get mentored by a human
-      %p On average, students iterate a further 3.5 times when mentored on a solution.
-      .c-combo-button
-        = link_to "Request Mentoring", "#", class: "--editor-segment"
-        %button.--dropdown-segment
-          = graphical_icon('chevron-down')
-
-      = link_to "#", class: 'learn-more' do
-        Learn more
-        = icon "external-link", "Opens in new tab"
